Move the remaining date-related functions out of `utils.py`
