+++ src/flickr_photos_api/utils.py
@@ -1,68 +1,6 @@
-import datetime
 from xml.etree import ElementTree as ET
 
-from .types import DateTaken, LocationInfo, SafetyLevel, Size, TakenGranularity
-
-
-def _parse_date_taken_value(p: str) -> datetime.datetime:
-    # See https://www.flickr.com/services/api/misc.dates.html
-    #
-    #     The date taken should always be displayed in the timezone
-    #     of the photo owner, which is to say, don't perform
-    #     any conversion on it.
-    #
-    # e.g. '2017-02-17 00:00:00'
-    return datetime.datetime.strptime(p, "%Y-%m-%d %H:%M:%S")
-
-
-def _parse_date_taken_granularity(g: str) -> TakenGranularity:
-    """
-    Converts a numeric granularity level in the Flickr API into a
-    human-readable value.
-
-    See https://www.flickr.com/services/api/misc.dates.html
-    """
-    lookup_table: dict[str, TakenGranularity] = {
-        "0": "second",
-        "4": "month",
-        "6": "year",
-        "8": "circa",
-    }
-
-    try:
-        return lookup_table[g]
-    except KeyError:
-        raise ValueError(f"Unrecognised date granularity: {g}")
-
-
-def parse_date_taken(
-    *, value: str, granularity: str, unknown: bool
-) -> DateTaken | None:
-    # Note: we intentionally omit sending any 'date taken' information
-    # to callers if it's unknown.
-    #
-    # There will be a value in the API response, but if the taken date
-    # is unknown, it's defaulted to the date the photo was posted.
-    # See https://www.flickr.com/services/api/misc.dates.html
-    #
-    # This value isn't helpful to callers, so we omit it.  This reduces
-    # the risk of somebody skipping the ``unknown`` parameter and using
-    # the value in the wrong place.
-    if unknown:
-        return None
-
-    # This is a weird value I've seen returned on some videos; I'm
-    # not sure what it means, but it's not something we can interpret
-    # as a valid date, so we treat "date taken" as unknown even if
-    # the API thinks it knows it.
-    elif value.startswith("0000-"):
-        return None
-
-    else:
-        return {
-            "value": _parse_date_taken_value(value),
-            "granularity": _parse_date_taken_granularity(granularity),
-        }
+from .types import LocationInfo, SafetyLevel, Size
 
 
 def parse_safety_level(s: str) -> SafetyLevel:
